Implementação de Rotas e Controle de Acesso

+++ app/Http/Controllers/ProdutoController.php
@@ -20,7 +20,7 @@ class ProdutoController extends Controller
                 'nome' => 'required|max:255',
                 'descricao' => 'required|max:255',
                 'valor' => 'required|regex:/^\d+(\.\d{1,2})?$/',
-                'avaliacao' => 'required|max:255',
+                
 
         ]);
 
@@ -36,27 +36,51 @@ class ProdutoController extends Controller
     }
 
  
-    public function update(Request $request, Produto $produto)
-    {
+
+    public function update(Request $request, $id)
+{
+    // Buscar o produto pelo ID
+    $produto = Produto::find($id);
+
+    // Se não encontrar, retorna erro
+    if (!$produto) {
+        return response()->json(['message' => 'Produto não encontrado.'], 404);
+    }
+
+    // Validação dos campos
         $fields =$request->validate([
             'nome' => 'required|max:255',
             'descricao' => 'required|max:255',
             'valor' => 'required|regex:/^\d+(\.\d{1,2})?$/',
-            'avaliacao' => 'required|max:255',
+            
 
     ]);
 
+    // Atualizar os campos do produto
     $produto->update($fields);
 
-        return [$produto];  
+    return response()->json([
+        'message' => 'Produto atualizado com sucesso!',
+        'produto' => $produto
+    ]);
 
     }
 
 
-    public function destroy(Produto $produto)
+    public function destroy($id)
     {
+        // Busca o produto pelo ID
+        $produto = Produto::find($id);
+    
+        // Verifica se o produto existe
+        if (!$produto) {
+            return response()->json(['error' => 'Produto não encontrado'], 404);
+        }
+    
+        // Deleta o produto
         $produto->delete();
-
-        return [ 'message' => 'Produto Excluído com Sucesso'];
+    
+        return response()->json(['message' => 'Produto excluído com sucesso'], 200);
     }
+    
 }
